math.ceil and math.floor should return integers, not floats.

+++ pallene/builtins.lua
@@ -24,8 +24,8 @@ builtins.modules = {
     },
     math = {
         abs   = T.Function({ T.Float() }, { T.Float() }),
-        ceil  = T.Function({ T.Float() }, { T.Float() }),
-        floor = T.Function({ T.Float() }, { T.Float() }),
+        ceil  = T.Function({ T.Float() }, { T.Integer() }),
+        floor = T.Function({ T.Float() }, { T.Integer() }),
         fmod  = T.Function({ T.Float(), T.Float() }, { T.Float() }),
         exp   = T.Function({ T.Float() }, { T.Float() }),
         ln    = T.Function({ T.Float() }, { T.Float() }),
